requirements.txt
MetaTrader5
pandas
numpy
python-dotenv
twilio